postgress/requirements.txt
psycopg2-binary 
pandas 
numpy
flask 
flask-cors